requirements.txt
# Core dependencies for PubMed medical literature search system
sentence-transformers>=2.2.0
lxml>=4.9.0
faiss-cpu>=1.7.0
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.24.0

# AI and ML dependencies
keybert>=0.7.0
anthropic>=0.7.0

# Data processing
xmltodict>=0.13.0
pyspark>=3.4.0

# Optional but recommended for better performance
torch>=2.0.0
transformers>=4.30.0

# Development and testing (optional)
pytest>=7.0.0
black>=23.0.0
